test: move checkpoint resume runtime tests
