+++ tests/test_runtime_orchestrator.py
@@ -302,236 +302,6 @@ def test_advance_execute_failure_keeps_previous_stage_state(tmp_path: Path) -> N
     assert not (task_dir / "execute-output.md").exists()
 
 
-def test_advance_rejects_mismatched_run_state_stage(tmp_path: Path) -> None:
-    policy = load_runtime_policy(ROOT)
-    task_dir = _make_task_dir(tmp_path)
-
-    run_state = json.loads((task_dir / "run-state.json").read_text(encoding="utf-8"))
-    run_state["current_stage"] = "execute"
-    _write_json(task_dir / "run-state.json", run_state)
-
-    with pytest.raises(RuntimeViolation, match="requested current_stage clarify does not match persisted run-state stage execute"):
-        advance_to_next_stage(task_dir=task_dir, policy=policy, route_name="small", current_stage="clarify")
-
-
-
-def test_resume_rejects_medium_session_state_with_wrong_plan_ledger_ref(tmp_path: Path) -> None:
-    policy = load_runtime_policy(ROOT)
-    task_dir = _make_task_dir(tmp_path)
-    _write_medium_plan_artifacts(task_dir, route_name="medium")
-    _write_json(
-        task_dir / "checkpoint.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "medium",
-            "current_stage": "clarify",
-            "plan_ref": "plan.json",
-            "plan_ledger_ref": "plan-ledger.json",
-            "run_state_ref": "run-state.json",
-            "next_action": "execute로 진행",
-            "open_blockers": [],
-            "updated_at": "2026-04-22T00:05:00Z"
-        },
-    )
-    _write_json(
-        task_dir / "session-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "medium",
-            "current_stage": "clarify",
-            "plan_ref": "plan.json",
-            "plan_ledger_ref": "run-state.json",
-            "run_state_ref": "run-state.json",
-            "latest_checkpoint_ref": "checkpoint.json",
-            "next_action": "execute로 진행",
-            "updated_at": "2026-04-22T00:05:00Z"
-        },
-    )
-
-    with pytest.raises(RuntimeViolation, match="session-state.json plan_ledger_ref run-state.json must point to plan-ledger.json for route medium"):
-        resume_task(task_dir=task_dir, policy=policy, route_name="medium")
-
-
-def test_resume_rejects_stale_session_state_latest_review_ref(tmp_path: Path) -> None:
-    policy = load_runtime_policy(ROOT)
-    task_dir = _make_task_dir(tmp_path)
-    _write_json(
-        task_dir / "review-report-spec.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "review_type": "spec",
-            "verdict": "approved",
-            "findings": ["spec ok"],
-            "approved_by": "spec-reviewer",
-            "next_action": "quality-review로 진행",
-        },
-    )
-    _write_json(
-        task_dir / "review-report-quality.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "review_type": "quality",
-            "verdict": "approved",
-            "findings": ["quality ok"],
-            "approved_by": "quality-reviewer",
-            "next_action": "finalize 가능",
-        },
-    )
-    _write_json(
-        task_dir / "checkpoint.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "small",
-            "current_stage": "clarify",
-            "plan_ref": "brief.json",
-            "plan_ledger_ref": "run-state.json",
-            "run_state_ref": "run-state.json",
-            "latest_review_ref": "review-report-quality.json",
-            "next_action": "execute로 진행",
-            "open_blockers": [],
-            "updated_at": "2026-04-22T00:05:00Z"
-        },
-    )
-    _write_json(
-        task_dir / "session-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "small",
-            "current_stage": "clarify",
-            "plan_ref": "brief.json",
-            "plan_ledger_ref": "run-state.json",
-            "run_state_ref": "run-state.json",
-            "latest_checkpoint_ref": "checkpoint.json",
-            "latest_review_ref": "review-report-spec.json",
-            "next_action": "execute로 진행",
-            "updated_at": "2026-04-22T00:05:00Z"
-        },
-    )
-
-    with pytest.raises(RuntimeViolation, match="session-state.json latest_review_ref review-report-spec.json does not match canonical latest review review-report-quality.json"):
-        resume_task(task_dir=task_dir, policy=policy, route_name="small")
-
-
-def test_status_rejects_session_state_route_drift(tmp_path: Path) -> None:
-    policy = load_runtime_policy(ROOT)
-    task_dir = _make_task_dir(tmp_path)
-    _write_json(
-        task_dir / "checkpoint.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "small",
-            "current_stage": "clarify",
-            "plan_ref": "brief.json",
-            "plan_ledger_ref": "run-state.json",
-            "run_state_ref": "run-state.json",
-            "next_action": "execute로 진행",
-            "open_blockers": [],
-            "updated_at": "2026-04-22T00:05:00Z"
-        },
-    )
-    _write_json(
-        task_dir / "session-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "medium",
-            "current_stage": "clarify",
-            "plan_ref": "brief.json",
-            "plan_ledger_ref": "run-state.json",
-            "run_state_ref": "run-state.json",
-            "latest_checkpoint_ref": "checkpoint.json",
-            "next_action": "execute로 진행",
-            "updated_at": "2026-04-22T00:05:00Z"
-        },
-    )
-
-    with pytest.raises(RuntimeViolation, match="session-state.json route medium does not match canonical route small"):
-        status_summary(task_dir=task_dir, policy=policy, route_name="small")
-
-
-def test_status_rejects_session_state_stage_drift(tmp_path: Path) -> None:
-    policy = load_runtime_policy(ROOT)
-    task_dir = _make_task_dir(tmp_path)
-    _write_json(
-        task_dir / "checkpoint.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "small",
-            "current_stage": "clarify",
-            "plan_ref": "brief.json",
-            "plan_ledger_ref": "run-state.json",
-            "run_state_ref": "run-state.json",
-            "next_action": "execute로 진행",
-            "open_blockers": [],
-            "updated_at": "2026-04-22T00:05:00Z"
-        },
-    )
-    _write_json(
-        task_dir / "session-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "small",
-            "current_stage": "execute",
-            "plan_ref": "brief.json",
-            "plan_ledger_ref": "run-state.json",
-            "run_state_ref": "run-state.json",
-            "latest_checkpoint_ref": "checkpoint.json",
-            "next_action": "execute로 진행",
-            "updated_at": "2026-04-22T00:05:00Z"
-        },
-    )
-
-    with pytest.raises(RuntimeViolation, match="session-state.json current_stage execute does not match run-state current_stage clarify"):
-        status_summary(task_dir=task_dir, policy=policy, route_name="small")
-
-
-def test_status_rejects_session_state_latest_review_ref_that_escapes_task_dir(tmp_path: Path) -> None:
-    policy = load_runtime_policy(ROOT)
-    task_dir = _make_task_dir(tmp_path)
-    _write_json(
-        task_dir / "checkpoint.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "small",
-            "current_stage": "clarify",
-            "plan_ref": "brief.json",
-            "plan_ledger_ref": "run-state.json",
-            "run_state_ref": "run-state.json",
-            "next_action": "execute로 진행",
-            "open_blockers": [],
-            "updated_at": "2026-04-22T00:05:00Z"
-        },
-    )
-    _write_json(
-        task_dir / "session-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "small",
-            "current_stage": "clarify",
-            "plan_ref": "brief.json",
-            "plan_ledger_ref": "run-state.json",
-            "run_state_ref": "run-state.json",
-            "latest_checkpoint_ref": "checkpoint.json",
-            "latest_review_ref": "../review-report.json",
-            "next_action": "execute로 진행",
-            "updated_at": "2026-04-22T00:05:00Z"
-        },
-    )
-
-    with pytest.raises(RuntimeViolation, match="session-state.json latest_review_ref ../review-report.json escapes task directory"):
-        status_summary(task_dir=task_dir, policy=policy, route_name="small")
-
 
 def test_run_route_rejects_medium_plan_ledger_out_of_order_completed_stages(tmp_path: Path) -> None:
     policy = load_runtime_policy(ROOT)
